fix LocalPath tests
